chore: added some additional helper files
