+++ src/helper/index.ts
@@ -16,201 +16,10 @@ import type {
     LeapDay,
     Moon
 } from "src/@types";
+import { DayHelper } from "./day.helper";
+import { MonthHelper } from "./month.helper";
 
-export class DayHelper {
-    private _events: FcEvent[];
-    shouldUpdate: boolean = false;
-    get calendar() {
-        return this.month.calendar;
-    }
-    get date() {
-        return {
-            day: this.number,
-            month: this.month.number,
-            year: this.year
-        };
-    }
-    get events(): FcEvent[] {
-        if (!this._events || !this._events.length || this.shouldUpdate) {
-            this._events = this.month.getEventsOnDay(this.date);
-        }
-        return this._events;
-    }
-    get longDate() {
-        return {
-            day: this.number,
-            month: this.month.name,
-            year: this.year
-        };
-    }
-    /** Days before this day in the year. */
-    get daysBefore() {
-        return (
-            this.month.daysBefore +
-            this.number -
-            1 -
-            this.month.leapDays.filter(
-                (l) => l.numbered && l.after < this.number - 1
-            ).length
-        );
-    }
-    get year() {
-        return this.month.year;
-    }
-    get weekday() {
-        const firstOfYear = this.calendar.firstDayOfYear(this.year);
-        return wrap(
-            (this.daysBefore % this.calendar.weekdays.length) + firstOfYear,
-            this.calendar.weekdays.length
-        );
-    }
-    get isCurrentDay() {
-        return (
-            this.number == this.calendar.current.day &&
-            this.month.number == this.calendar.current.month &&
-            this.month.year == this.calendar.current.year
-        );
-    }
-    get isDisplaying() {
-        return (
-            this.number == this.calendar.viewing.day &&
-            this.month.year == this.calendar.viewing.year &&
-            this.month.number == this.calendar.viewing.month
-        );
-    }
-    private _moons: Array<[Moon, Phase]>;
-    get moons() {
-        if (!this._moons || !this._moons.length) {
-            this._moons = this.month.getMoonsForDay(this.date);
-        }
-        return this._moons;
-    }
 
-    constructor(
-        public month: MonthHelper,
-        public number: number,
-        public leapday?: LeapDay
-    ) {}
-}
-
-export class MonthHelper {
-    days: DayHelper[] = [];
-    daysBefore: number;
-    leapDays: LeapDay[] = [];
-    shouldUpdate = false;
-
-    get id() {
-        return this.data.id;
-    }
-    get index() {
-        return this.calendar.data.months.indexOf(this.data);
-    }
-    get name() {
-        return this.data.name;
-    }
-    get length() {
-        return this.days.length;
-    }
-
-    get firstWeekday() {
-        if (!this.calendar.data.overflow) return 0;
-        return this.days[0].weekday;
-    }
-    get lastWeekday() {
-        return this.days[this.days.length - 1].weekday;
-    }
-
-    get type() {
-        return this.data.type;
-    }
-    events: FcEvent[];
-    getEventsOnDay(day: CurrentCalendarData) {
-        if (!this.events || !this.events.length || this.shouldUpdate) {
-            this.days.forEach((day) => (day.shouldUpdate = true));
-            this.events = this.calendar.getEventsForMonth(this);
-            this.shouldUpdate = false;
-        }
-        return this.events.filter((event) => {
-            if (
-                (!event.date.year || event.date.year == day.year) &&
-                (!event.date.month || event.date.month == day.month) &&
-                event.date.day == day.day
-            )
-                return true;
-            if (!event.end && !event.formulas?.length) return false;
-            const start = { ...event.date };
-            const end = {
-                ...(event.end ?? {})
-            };
-
-            if (!start.year) start.year = end.year = this.year;
-            if (!start.month) start.month = end.month = this.number;
-
-            const hash = Number(this.calendar.hash(day));
-            if (
-                Number(this.calendar.hash(start)) <= hash &&
-                hash <= Number(this.calendar.hash(end) ?? Infinity)
-            ) {
-                if (!event.formulas?.length) {
-                    return true;
-                } else {
-                    const startDays =
-                        this.calendar.totalDaysBeforeYear(start.year) +
-                        this.calendar.daysBeforeMonth(
-                            start.month,
-                            start.year,
-                            true
-                        ) +
-                        start.day;
-                    const currentDays =
-                        this.calendar.totalDaysBeforeYear(day.year) +
-                        this.calendar.daysBeforeMonth(
-                            day.month,
-                            day.year,
-                            true
-                        ) +
-                        day.day;
-                    const daysBetween = currentDays - startDays;
-
-                    return daysBetween % event.formulas[0].number == 0;
-                }
-            }
-            return false;
-        });
-    }
-    shouldUpdateMoons = false;
-    moons: Array<[Moon, Phase]>[];
-    getMoonsForDay(day: CurrentCalendarData) {
-        if (!this.moons || !this.moons.length || this.shouldUpdateMoons) {
-            this.moons = this.calendar.getMoonsForMonth(this);
-        }
-        return this.moons[day.day - 1];
-    }
-    constructor(
-        public data: Month,
-        public number: number,
-        public year: number,
-        public calendar: CalendarHelper
-    ) {
-        this.leapDays = this.calendar.leapDaysForMonth(this.number, year);
-        this.daysBefore = this.calendar.daysBeforeMonth(this.number, this.year);
-
-        this.days = [
-            ...new Array(
-                data.length +
-                    this.leapDays.filter(
-                        (l) => !l.intercalary || (l.intercalary && l.numbered)
-                    ).length
-            ).keys()
-        ].map((k) => {
-            return new DayHelper(
-                this,
-                k + 1,
-                this.leapDays.find((leapday) => leapday.after == k)
-            );
-        });
-    }
-}
 
 interface YearEventCache {
     events: FcEvent[];
